requirements.txt
flask
msrest
azure-storage-blob
azure-cognitiveservices-vision-computervision
werkzeug
flask-cors
